pyproject.toml
[build-system]
requires = ["hatchling", "hatch-cython"]
build-backend = "hatchling.build"

[project]
name = "Pluggable Namespaces"
version = "1.0.0"
description = "Pluggable namespaces"
readme = "README.rst"
authors = [
    {name = "Tyler Levy Conde", email = "yonstib@gmail.com"},
]
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License",
]
requires-python = ">=3.11"
dependencies = [
    # Required for logging to a file
    "aiofiles",
    "aiologger",
    "argparse",
    "msgpack",
    "PyYaml",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
]

build = [
    "Cython",
    "hatch",
    "hatch-cython",
]

[tool.hatch.build.targets.wheel]
packages = [
    "src/cpns",
    "src/pns",
]

[tool.hatch.build.hooks.convert-data-file.options]
pattern = "config.yaml"
target_format = "msgpack"

[tool.hatch.build.hooks.cython.options]
include_numpy = false
include_pyarrow = false
include_pythran = false
directives = { boundscheck = false, nonecheck = false, language_level = 3, binding = true, cdivision = true, wraparound=false }
compile_args = [
  "-v",
  "-O3",
  "-flto"
]

[tool.pytest.ini_options]
testpaths = "tests"
addopts = "--tb native --full-trace --color=yes -vv"
asyncio_mode = "auto"
